// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const Role = {
	LEARNER: "LEARNER",
	MENTOR: "MENTOR",
	ADMIN: "ADMIN",
	SUPER_ADMIN: "SUPER_ADMIN",
} as const;

export type Role = (typeof Role)[keyof typeof Role];

export const Status = {
	ACTIVE: "ACTIVE",
	BLOCKED: "BLOCKED",
	DELETED: "DELETED",
} as const;

export type Status = (typeof Status)[keyof typeof Status];

export const Gender = {
	MALE: "MALE",
	FEMALE: "FEMALE",
	OTHER: "OTHER",
} as const;

export type Gender = (typeof Gender)[keyof typeof Gender];